Refactored project structure
Type: trivial
