+++ pgtrigger/__init__.py
@@ -1,43 +1,48 @@
 import django
 
+from pgtrigger.api import (
+    constraints,
+    disable,
+    enable,
+    get,
+    ignore,
+    install,
+    prunable,
+    prune,
+    register,
+    schema,
+    uninstall,
+)
+from pgtrigger.contrib import (
+    FSM,
+    Protect,
+    SoftDelete,
+    UpdateSearchVector,
+)
 from pgtrigger.core import (
     After,
     Before,
     Condition,
-    constraints,
     Deferred,
     Delete,
-    disable,
-    enable,
     F,
-    FSM,
-    get,
-    ignore,
     Immediate,
     Insert,
-    install,
     InsteadOf,
     IsDistinctFrom,
     IsNotDistinctFrom,
     Level,
     Operation,
     Operations,
-    Protect,
-    prune,
     Q,
     Referencing,
-    register,
     Row,
-    schema,
-    SoftDelete,
     Statement,
     Timing,
     Trigger,
     Truncate,
-    uninstall,
     Update,
     UpdateOf,
-    UpdateSearchVector,
     When,
 )
 from pgtrigger.version import __version__
@@ -71,6 +76,7 @@ __all__ = [
     'Operation',
     'Operations',
     'Protect',
+    'prunable',
     'prune',
     'Q',
     'Referencing',
